% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIC.R
\name{checkIC}
\alias{checkIC}
\title{Check for cointegration, unlikely initial transients}
\usage{
checkIC(vout, T = NULL, divider = NULL, ic = NULL, icx = 1)
}
\arguments{
\item{vout}{A list like that produced as output from \code{rfmdd()}.  At least the
\code{var$By} componnent must be present.  With \code{T} and \code{divider} both
\code{NULL}, the \code{uts} component is needed.  To get useful results
about initial transients, the \code{uts} and \code{var$Bx}  components are
needed.}

\item{T}{Roots closer than \code{1/T} to 1.0 (in the complex plane) are treated as
non-stationary.  Such roots can imply deterministic trend-like
components in the data.}

\item{divider}{If this is non-null and \code{T} is null, roots larger in absolute
value than \verb{divider are treated as non-stationary. }divider > 1` precludes transient check.}

\item{ic}{Initial conditions. \code{lags} by \code{ny} matrix of initial data. Needed
for check vs large initial transients.}

\item{icx}{Mean values for x variables.  Default of 1 is right if the only x
is the constant.}
}
\value{
\describe{
\item{cointvecs}{\code{ny - m} by \code{ny} matrix. If \code{m  < ny} eigenvalues are
big, as determined by \code{T} and \code{divider}, the cointegrating
vectors (linear combinations of \code{y} that grow more slowly
than the \code{m} large-root components) the rows of cointmat.}
\item{z0tstat}{A vector of ratios of deviations of initial condition
values of the stationary components of \code{y} from their means,
to their implied uncondiional standard errors.  If any
of these are large (e.g., > 3.5), the model implies
initial transients that are unlikely to occur again.}
\item{z0chisq}{A chi-squared quantity analogous to the individual elements
of \code{z0tstat}}
\item{roots}{eigenvalues of the system, ordered as in \code{z0tstat}.}
}
}
\description{
Splits eigenvalues of a VAR system into large and small categories.  If
the system has n variables and m < n eigenvalues are large, calculates
coefficients of cointegrating vector. Checks for initial conditions implying
large and rare transients.
}
\details{
By default, it treats as large all eigenvalues that differ from one (in the
complex plane) by less than \code{1/T}, where \code{T} is sample size, or else excced
one in absolute value.  \code{T} can optionally be set by the user.  If \code{T} is
\verb{NULL' and }divider\verb{is not, roots greater than}divider` in absolute value
are treated as large.
}
