% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubling.R
\name{doubling}
\alias{doubling}
\title{Doubling algorithm}
\usage{
doubling(A, omega, crit = 1e-09)
}
\arguments{
\item{A}{The matrix of AR coefficients.}

\item{omega}{The covariance matrix of disturbances}

\item{crit}{When the algorithm changes the result by less than this, it stops.}
}
\value{
The unconditional covariance matrix \code{V}.
}
\description{
Calculates unconditional covariance for a stationary, first-order
VAR system.
}
\seealso{
\code{sysmat}, which takes the \code{var$By}, or \code{var$By}
and \code{var$Bx}, components of \code{rfmdd} output  and forms the \code{A}
matrix for a stacked 1st order system.
}
