% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcast.R
\name{fcast}
\alias{fcast}
\title{Forecast}
\usage{
fcast(y0, By, Bx, xdata = NULL, const = TRUE, horiz, shocks = NULL)
}
\arguments{
\item{y0}{lags x nvar matrix (even if 1-d) of initial conditions}

\item{By}{equations x variables x lags array of AR coefficients}

\item{Bx}{neq x nx matrix of coefficients on exogenous variables}

\item{xdata}{(lags + horiz) x nx matrix of exogenous variable values}

\item{const}{TRUE if model contains constants}

\item{horiz}{number of periods over which forecast will be made}

\item{shocks}{horiz x nvar matrix of shocks during the forecast period}
}
\value{
(lags + horiz) x nvar matrix of initial conditions and forecast values.
}
\description{
Forecast with a reduced form VAR
}
