% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcastCndl.R
\name{fcastCndl}
\alias{fcastCndl}
\title{Conditional forecast}
\usage{
fcastCndl(
  y0,
  Ay,
  Ax,
  xdata = NULL,
  const = TRUE,
  horiz,
  R = NULL,
  g = NULL,
  whichShocks = rep(TRUE, dim(y0)[2]),
  yr = NULL
)
}
\description{
Forecast conditional on some future \code{y} values and on zero values for some
shocks.
}
\details{
Model is an SVAR: \eqn{Ay(L)y = Ax(L)x + \eps} with \eqn{\eps ~ N(0,I)}.
The conditioning information is \code{R \%*\% y = g}, where \code{y} is the \code{(lags + horiz)}
by \code{nv} data and forecast matrix, stacked into a single vector.

The conditional forecast will be generated using the shocks for which
\code{whichShocks == TRUE}.  Alternatively, a horiz x nv matrix can be
provided witn NaN values in all positions except those at values of
$y$ that are constrained, and the constrained values in other positions.
Note that if a reduced form VAR y=By(L)y+e is being used, Ay(L) is I - By(L)
only if Var(e)=I.  If Var(e)=Sigma in the reduced form,
A(L) = S \%\emph{\% (I-B(L)), where S \%}\% Sigma \%*\% t(S) = I. Of course the choice
of S affects the result in this case.
}
