% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcastMany.R
\name{fcastMany}
\alias{fcastMany}
\title{Many k-step-ahead forecasts}
\usage{
fcastMany(ydata, By, Bx, xdata = NULL, const = TRUE, horiz)
}
\description{
Calculates forecasts and forecast errors at horizons given by the horiz vector
at each date from lags+1 through end(ydata).  \code{fcast()}
xdata, if present, must have first dimension exceeding that of ydata by max(horiz)
Parameter descriptions still needed.
}
