% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulsdtrf.R
\name{impulsdtrf}
\alias{impulsdtrf}
\title{Impulse responses for VAR}
\usage{
impulsdtrf(vout = NULL, smat = NULL, nstep = 40, order = NULL)
}
\arguments{
\item{vout}{Output from VAR estimation, \code{rfmdd()} or \code{svmdd} or the \verb{$vout}
element from \code{svarwrap()} invoked with \code{verbose=TRUE}.}

\item{smat}{If order and smat are NULL, the impulse responses will be for
a cholesky decomp with variables ordered as in \code{vout}.
More generally, smat can be any set of initial values for the
shocks. For an SVAR, 'smat=solve(A0)'.}

\item{nstep}{Number of time steps over which to compute the impulse response.}

\item{order}{To get a cholesky decomp with a different ordering, set order
to an integer vector giving the desired ordering.}
}
\value{
\code{nvar} x \code{nshocks} x \code{nstep} array of impulse responses.
}
\description{
Uses output in \code{rfvar} format to construct impulse responses.
}
\details{
To use this with output from \code{postdraw}, create a dummy \code{vout} with
\code{vout$By=pout$By[ , , ,id]} and provide \code{smat=pout$smat[ , ,id]}.
To keep the shocks orthogonal and scaled properly, smat should be such that
\code{smat \%*\% t(smat) == cov(vout$u)}.  This means that the \code{smat} from
\code{postdraw()} has to be transposed.  The routine also works with
singular \code{smat} or with \code{smat} column dimension less than its row dimension.
}
