% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invTime.R
\name{invTime}
\alias{invTime}
\title{Inverse Time}
\usage{
invTime(tsdate, tsobj)
}
\arguments{
\item{tsdate}{Date from \code{ts} object}

\item{tsobj}{Object on whose dating the calculation is based.}
}
\description{
Converts \code{ts} object dates to integer indexes into the \code{ts} object
}
\details{
\code{tsdate} can either be a vector of real numbers, e.g. 1973.25
or \code{c(1973.25, 1990+2/3)}, or a matrix with 2 columns containing
dates in year-month pairs \verb{c(1973,2), c(1990,9)} (for monthly data, e.g.).
If the date is outside the start, end range, the returned observation
number will be zero or N+1 where N is the number of rows in tsobj.
To avoid ambiguity, A single date of the form \code{c(1990,2)} must be
dimensioned as a 1 by 2 matrix.
}
