% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lu.R
\name{lu}
\alias{lu}
\title{Simple LU decomposition}
\usage{
lu(x)
}
\arguments{
\item{x}{Matrix to be decomposed}
}
\value{
List with components \code{L}, \code{U} and rowperm. \code{L} lower triangular, \code{U} upper
triangular, `L \%*\% U == x". rowperm shows possible reordering of rows.
}
\description{
Calculates LU decomposition with row pivoting.
}
\details{
No advantage for this over the Matrix package function of the same name,
except that this handles complex matrices and possibly that this works
with only the standard dense matrix data type.
}
