% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnpparamSet.R
\name{mnpparamSet}
\alias{mnpparamSet}
\title{Set VAR prior parameters}
\usage{
mnpparamSet(
  nv = 1,
  nx = 1,
  lags = 3,
  tight = 4,
  decay = 0.3,
  sig = rep(0.01, nv),
  w = 1,
  lambda = 5,
  mu = 1,
  xsig = NULL,
  ybar = NULL,
  xbar = 1,
  OwnLagMeans = c(1.25, -0.25)
)
}
\description{
Returns a default list of parameters, or allows varying elements of the list.
}
\details{
This could be useful in documenting results.  If the list returned from this
program (call it \verb{mnplist') is used in }with(mnplist, varprior(nx, lags, tight, decay, etc])',
you have the actual, evaluated arguments in 'mnplist'.

Probably better, as it avoids writing out the whole arg list each call, is
\code{do.call(varprior, mnplist)}
}
