% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pchol.R
\name{pchol}
\alias{pchol}
\title{Permuted Cholesky decomposition}
\usage{
pchol(sig, porder)
}
\arguments{
\item{sig}{Matrix to factor.}

\item{porder}{Permutation vector to apply to rows and columns}
}
\value{
The matrix rows and columns are permuted, the factorization
is done, and the permutation is then undone.  So the returned value
is a matrix square root of \code{sig} (i.e. satisfies \code{crossprod(w)==sig})
but not triangular.
}
\description{
Permuted Cholesky decomposition
}
