% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfc.R
\name{plotfc}
\alias{plotfc}
\title{Forecast plots with error bands}
\usage{
plotfc(
  fc,
  tspfc,
  var = NULL,
  type = c("multiple", "single"),
  main = "Forecast plot",
  yax.flip = TRUE,
  file = "fcplot.pdf"
)
}
\arguments{
\item{fc}{3-d array of forecasts.}

\item{tspfc}{time series tsp attribute of the forecasts (start, end, frequency)}

\item{var}{Indexes of variables to be plotted, if not all of them.}

\item{type}{"multiple" creates an array of plots. "single" puts them
all on same plot (usually responses of just one variable).
(Curent version ignores this argument, always assumes "multiple".)}

\item{main}{Title for plot.}

\item{yax.flip}{Alternate sides of small plots on which y-axis labels appear?}

\item{file}{Name for output pdf file.}
}
\description{
Plots an array of forecast graphs
}
