% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postdraw.R
\name{postdraw}
\alias{postdraw}
\title{Draw from VAR posterior}
\usage{
postdraw(vout, n, nosigprior = FALSE)
}
\arguments{
\item{vout}{Output from \code{rfvar}, or a list with the same names, like \code{rfmdd()}
output lisst's  \verb{$var} element.}

\item{n}{Number of draws.}

\item{nosigprior}{If TRUE, don't use the Jeffreys-motivated improper prior.}
}
\value{
\item{By}{Draws of AR coefficient array.}
\item{Bx}{Draws of constant terms or exogenous variable coefficients}
\item{smat}{\code{chol(sigma(draw))}.  Transpose to use as smat in \code{impulsdtrf}}
}
\description{
Draw from the joint posterior distribution of a vector autoregression's coefficients and
residual covariance matrix.
}
