% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmdd.R
\name{rfmdd}
\alias{rfmdd}
\title{Reduced form VAR estimation}
\usage{
rfmdd(
  ydata,
  lags,
  xdata = NULL,
  const = TRUE,
  lambda = 5,
  mu = 1,
  tight = 3,
  decay = 0.5,
  sig = rep(0.01, dim(ydata)[2]),
  w = 1,
  xsig = NULL,
  OwnLagMeans = c(1.25, -0.25),
  flat = FALSE,
  nonorm = FALSE,
  ic = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ydata}{Endogenous variable data matrix, including initial condition
dates.  Usually just an mts object.  More generally, it may
be a list of mts objects that will be stacked up for estimation.}

\item{lags}{Number of lags.}

\item{xdata}{Exogenous variable data matrix, including initial condition
dates. A list when ydata is a list.}

\item{const}{Create constant term (with no need for column of ones in
\code{xdata})}

\item{lambda}{Weight on the co-persistence prior dummy observation.  If
negative, does not include x's in the dummy observation.}

\item{mu}{Weight on variable-by-variable sum of coeffs dummy observations.
if negative, does not include x's in the dummy observations}

\item{tight}{Overall tightness of prior.  Larger is tighter. Prior standard
error of own-lag coefficient is 1/tight}

\item{decay}{Prior standard deviations of coefficients decrease as
\code{1/lag^decay}.}

\item{sig}{Modal prior value for standard deviationso of residuals.  A
vector of length nv.  Even if \code{w == 0}, this vector is needed
for scaling other parts of the prior.}

\item{w}{weight on prior dummy observations pulling residual standard
deviations toward \code{sig}.}

\item{xsig}{Scales of variation in \code{x} variables.  If non-NULL, the last
element matches the constant and should usually be zero.}

\item{OwnLagMeans}{Prior expectation of own lag coefficients.  See details.}

\item{flat}{Omit conventional uninformative prior on \code{Sigma}?}

\item{nonorm}{Do not normalize posterior to make it a proper prior}

\item{ic}{If non-null, do not use initial conditions from \code{ydata} in
forming the prior.  Use \code{ic} instead.}

\item{verbose}{If true, return real-data residuals with correct dating
and prior parameters.}
}
\value{
\item{mdd}{Log of integrated posterior}
\item{var}{\code{rfvar} return list using all observations,
including dummies}
\item{varp}{\code{rfvar} return list using only dummy observations.}
Items below returned only when \code{verbose==TRUE}.
\item{uts}{residuals corresponding to data (not dummy obs), either
as a single mts object or as a list of such objects,
depending on whether \code{ydata} was mts or a list.}
\item{prior}{list of prior hyperparameter settings used}
\item{pintp}{Log of integrated density for dummy observations; scale
factor to convert them to proper prior.}
\item{call}{The function call invoking this function that produced
this result.}
}
\description{
Estimate a reduced form VAR and (optionally) find its integrated posterior
(mdd)
}
\details{
The marginal density results are only available if the prior is proper, which
usually requires all or most of the prior parameters to be non-zero.  But
estimation with an improper prior that uses only some of the Minnesota prior
dummy observations is also possible.  For example \code{lambda=3}, \code{mu=1}, \code{tight=0},
\code{w=0},  (but \code{sig} non-zero) is a loose improper prior that insulates
against estimates that imply unlikely initial transients.   When the prior is
improper, \code{nonorm=TRUE} avoids some unnecessary calculations.

For more extensive discussion of the prior parameters see \link{MNpriorNotes.pdf}.

Though most commonly \code{ydata} is just a \code{mts} object, it can also be a list
of such objects.  The main use of this is to allow use of time series with
missing observations.  It also allows direct entry of dummy observations
to function as part of the prior.

Implementing a conjugate prior not in the parametric class allowed for in
\code{varprior()} is possible by specifying \code{ydata} as a list and giving one or
more of the elements of the list an attribute named \code{dummy} and set to TRUE.
The blocks labeled this way are treated as part of the prior.  Changing or
eliminating the \code{dummy} attribute does not change the posterior distribution,
but it does affect the value of \code{mdd} and thus Bayes factors for model
comparison.

To make the first T0 observations a pure "training sample", make \code{ydata} a
list with the first component the data from 1 to T0 + lags and the second
component the data from T0 to the end of the sample.

\code{OwnLagMeans} may be a single numeric value, the prior mean of the
first own lag coefficient in all equations.  If it is a numeric vector of
length m, the prior mean of the first m lag coefficients in all equations.
If it is a \code{nv} by \code{m} matrix, each row is the prior mean of the coefficients
on the first \code{m} lags in the corresponding equation.

The default is close to the optimal second-order univariate AR coefficients
when the variable is a unit-averaged continuous time Wiener process.  This
works well for variables like GDP or investment, which cumulate through time.
For data that are sampled rather than averaged, like some financial or price
data, \code{OwnLagMeans=1} is better.
}
