% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfvar.R
\name{rfvar}
\alias{rfvar}
\title{Reduced form VAR estimation}
\usage{
rfvar(ydata = NULL, lags = 6, xdata = NULL, breaks = NULL)
}
\arguments{
\item{ydata}{\verb{T x nvar} dependent variable data matrix.}

\item{lags}{number of lags in the VAR}

\item{xdata}{\verb{T x nx} exogenous variable data matrix.}

\item{breaks}{Vector of row numbers in \code{ydata} and \code{xdata} after which
there is a break, including T.}
}
\value{
\describe{
\item{By}{nvar x nvar x lags array of coefficients on lagged ys.
1st dimension is equation number}
\item{Bx}{nvar x nx matrix of coefficients on x}
\item{u}{Residuals. Note that This matrix will have fewer rows than
\code{ydata} because of the \code{breaks}.  Residuals at the end
may be from dummy observations.}
\item{xxi}{\code{crossprod(X)} \code{kronecker(cov(u),xxi)} is the full
covariance matrix of the regression coefficients.}
\item{snglty}{ Usually 0.  If the rhs variable matrix is not full column
rank, this is the gap between the number of columns and the
number of non-zero singular values.
}
}
}
\description{
Estimates a vector autoregression, using a prior expressed as dummy variable
observations.
}
\details{
This program does not set a prior distribution, calculate an integrated
posterior (marginal data density), or distinguish dummy observations from
data. \code{rfmdd()}, which can do all that, uses this program.

The prior is implemented outside this program with dummy observations, which
are included, after the "real" data, at the end of \code{ydata}.  The \code{breaks}
vector specifies rows of \code{ydata} after which the next lags rows are initial
conditions for the next batch of data. Dummy observations for a prior
are usually single blocks of lags+1 rows.  But \code{breaks} can also
be used to omit blocks of rows of \code{ydata}, or to specify where data for
a specific country ends if a single VAR is being fit to data for several
countries. \code{breaks} includes, as its last element, the number of rows in
\code{ydata}.
}
