% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsf2csf.R
\name{rsf2csf}
\alias{rsf2csf}
\title{Real Schur form to complex}
\usage{
rsf2csf(Q, T)
}
\arguments{
\item{Q}{The orthonormal matrix from the real Schur}

\item{T}{The nearly triangular matrix from the real Schur}
}
\description{
Real Schur form to complex
}
\details{
The real form is \eqn{A = QTQ'} with QQ'=I and T triangular, except
possibly for non-zero elements on the first diagonal below the main diagonal.
The complex form has T fully triangular, with Q' interpreted as both
conjugated and transposed.
}
