% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwwish.R
\name{rwwish}
\alias{rwwish}
\title{Wishart random matrices}
\usage{
rwwish(v, S, n)
}
\arguments{
\item{v}{degrees of freedom for the Wishart}

\item{S}{Scale matrix for the Wishart}

\item{n}{number of draws needed}
}
\value{
a p x p x n array of n draws that are Cholesky square roots of Wisharts.
}
\description{
This is much more efficient than the original MCMCpack rwish when multiple
draws are to be made.  about 20 times faster for 4x4 S with n=200, e.g.
Because this is based on
MCMCpack code, it is covered by the GPL, version 2, which is available with
every R installation.#'
}
