% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svar.R
\name{svar}
\alias{svar}
\title{SVAR estimation}
\usage{
svar(
  ydata = NA,
  lags = 6,
  xdata = NULL,
  const = TRUE,
  breaks = NULL,
  A0 = diag(NCOL(ydata)),
  Tsigbrk = NROW(ydata),
  lmd = matrix(1, NCOL(ydata), length(Tsigbrk)),
  nonorm = FALSE
)
}
\arguments{
\item{ydata}{Endogenous variable data matrix, including initial condition
dates.  Usually just an mts object.  More generally, it may
be a list of mts objects that will be stacked up for estimation.}

\item{lags}{number of lags in the model}

\item{xdata}{T x nx exogenous variable data matrix.}

\item{breaks}{Column vector of row numbers in ydata and xdata after which
there is a break in the time series.}

\item{A0}{nvar x nvar matrix of coefficients on contemporary variables.}

\item{Tsigbrk}{Each entry of this vector indexes the last row of \code{ydata} in
the corresponding variance regime.}

\item{lmd}{nvar x length(Tsigbrk) matrix of relative variances in regimes.
Each row normalized to average to one across the real data.}

\item{nonorm}{When \code{TRUE}, marginal data density is not calculated.
Can save a little time if mdd not needed, and is necessary if
dummy observations do not determine a proper prior.}
}
\value{
\describe{
\item{By}{nvar x nvar x lags matrix of coefficients on lagged \code{y}'s.
1st dimension is equation number}
\item{Bx}{nvar x nx matrix of coefficients on x}
\item{u}{\code{T-lags + (number of dummy obs)} by \code{nvar} matrix of
residuals.}
\item{uraw}{Unweighted reduced-form residuals --- one-step-ahead
prediction errors}
\item{xxi}{Cross-product of weighted \code{X} matrix.  One for each equation.
Equation \code{i}'s right-hand-side coefficient covariance matrix.}
\item{snglty}{ Usually 0.  If the rhs variable matrix is not full column
rank, this is the gap between the number of columns and the
number of non-zero singular values.}
}
}
\description{
Structural VAR estimation with time-varying shock variances
}
\details{
Works from a data matrix that includes any dummy observations implementing
a conjugate prior. Usually called from \code{svmdd()} rather than directly.
This program does not set a prior distribution:  It assumes the prior has
been implemented with dummy observations that are included in \code{ydata}.
\subsection{\code{breaks}}{The first \code{lags} rows of \code{ydata} after an
element of \code{breaks} are used as new initial conditions, not
left-hand-side predicted values. This allows for gaps in the time
series and for the possibility of adding dummy observations to
implement a prior.}
\subsection{\code{A0}}{constant matrix such that \code{A0 \% * \% u},
where \code{u} is the reduced form VAR residual, has a diagonal covariance
matrix.}
\subsection{Varying variances}{ The rows of \code{lmd} are constrained to average
to 1, and rows in ydata after the last element of \code{Tsigbrk}, which should
all be prior dummy observations, are given a vector of ones as \code{lmd} entries.}
}
\seealso{
\code{\link[=svmdd]{svmdd()}}
}
