% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svarwrap.R
\name{svarwrap}
\alias{svarwrap}
\title{log posterior density for SVAR with time series ID through heteroskedasticty}
\usage{
svarwrap(
  x,
  ydata = NULL,
  lags = 5,
  xdata = NULL,
  const = TRUE,
  Tsigbrk = NULL,
  tight = 1,
  decay = 0.3,
  lambda = 5,
  mu = 1,
  sig = rep(0.01, NCOL(ydata)),
  alpha = 6,
  OwnLagMeans = c(1.25, -0.25),
  verbose = FALSE
)
}
\arguments{
\item{x}{Vector containing A0 and all but the last column of lmd, vectorized.}

\item{ydata}{Endogenous variable data matrix, including initial condition
dates.  Usually just an mts object.  More generally, it may
be a list of mts objects that will be stacked up for estimation.}

\item{lags}{Number of lags in the model.}

\item{xdata}{Exogenous variable data matrix, including initial condition
dates. A list when ydata is a list.}

\item{const}{Constant term is added automatically if const=TRUE.}

\item{Tsigbrk}{Dates at which lmd vectors change.  Last date with old lmd (not first
with new).  Can be two-column matrix with rows like c(1947,2),
or vector with absolute dates like 1947.25.}

\item{tight}{Overall tightness of Minnesota prior. \code{1/tight} is own lag std dev}

\item{decay}{Standard deviations of lags shrink as \code{lag^(-decay)}.
walk behavior}

\item{lambda}{Weight on the co-persistence prior dummy observation.  If
negative, does not include x's in the dummy observation.}

\item{mu}{Weight on variable-by-variable sum of coeffs dummy observations.
if negative, does not include x's in the dummy observations}

\item{sig}{Vector of prior guesses at standard deviations of reduced form
residuals.}

\item{alpha}{Parameter of symmetric Dirichlet distribution for the elements
of each row of the lmd matrix.  Increasing alpha increases the
prior's shrinkage toward equal variances in sub-periods.}

\item{OwnLagMeans}{Prior expectation of own lag coefficients in reduced form.
See details.}

\item{verbose}{If FALSE, return only the log marginal posterior density for `A0,lmd'.}

\item{flat}{Omit conventional uninformative prior on \code{Sigma}?}

\item{nonorm}{Do not normalize posterior to make it a proper prior}

\item{ic}{If non-null, do not use initial conditions from \code{ydata} in
forming the prior.  Use \code{ic} instead.}
}
\value{
\itemize{
\item \code{lh}: Minus log posterior density. If \code{verbose==FALSE}, only this is returned
\item \code{vout}: Output from \code{svmdd}.  This includes the input prior parameters,
reduced form residuals (variable innovations), and much else.  See
\code{\link[=svmdd]{svmdd()}}.
\item \code{A0}: Matrix of contemporaneous coefficients on \eqn{y}
\item \code{lmd}: Matrix of relative variances of residuals in each regime
\item \code{lplmd}: log density from lmd prior
\item \code{allh}: log density from A0 prior
}
}
\description{
Variances of structural shocks change at pre-specified break dates.
}
\details{
Parameter \code{A0} is the lead coefficient in \eqn{A(L)}, with the model
\eqn{A(L)y = \epsilon}.   \code{A0} has the identity matrix times \code{100 * asig}  as
its prior mean, implying residual std deviations centered at about
\code{.01 / asig}.  All elements of \code{A0} have std deviation \code{200 * asig}, making
the prior fairly loose. The program should give more control over the prior
on A0. In the meantime, ydata should be scaled to make the equal prior means
of shock scales plausible.

\code{lmd} is a matrix defining the relative structural shock variances in
the different periods defined by TsigBrk.  Each column of lmd corresponds to
one of these periods, and each row corresponds to one shock variance.  Each
row is normalized to average to one.  The priors on \code{lmd} are independent
across rows and are scaled Dirichlet(2).

For more detailed discussion of prior parameters see \code{\link[=svmdd]{svmdd()}}.
}
