% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmdd.R
\name{svmdd}
\alias{svmdd}
\title{Structural VAR estimation}
\usage{
svmdd(
  ydata,
  lags,
  xdata = NULL,
  const = TRUE,
  A0,
  Tsigbrk,
  lmd,
  lambda = 5,
  mu = 1,
  tight = 3,
  decay = 0.3,
  sig = rep(0.01, NCOL(ydata)),
  xsig = NULL,
  OwnLagMeans = c(1.25, -0.25),
  flat = FALSE,
  nonorm = FALSE,
  ic = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ydata}{Endogenous variable data matrix, including initial condition
dates.  Usually just an mts object.  More generally, it may
be a list of mts objects that will be stacked up for estimation.}

\item{lags}{Number of lags in the model.}

\item{xdata}{Exogenous variable data matrix, including initial condition
dates. A list when ydata is a list.}

\item{const}{Constant term is added automatically if const=TRUE.}

\item{A0}{Contemporaneous coefficient matrix --- constant.}

\item{Tsigbrk}{Dates at which lmd vectors change.  Last date with old lmd (not first
with new).  Can be two-column matrix with rows like c(1947,2),
or vector with absolute dates like 1947.25.}

\item{lmd}{Relative  variances of structural shocks.  Rows for variables,
columns for periods.  Row-averages normalized to one.}

\item{lambda}{Weight on the co-persistence prior dummy observation.  If
negative, does not include x's in the dummy observation.}

\item{mu}{Weight on variable-by-variable sum of coeffs dummy observations.
if negative, does not include x's in the dummy observations}

\item{tight}{Overall tightness of Minnesota prior. \code{1/tight} is own lag std dev}

\item{decay}{Standard deviations of lags shrink as \code{lag^(-decay)}.}

\item{sig}{Scales for prior on A0. Should be the order of magnitude of
the std deviations of first differences of the variables, usually.}

\item{xsig}{Scales for the prior on variation in the \code{x} variables. If there
is only a constant, this should be \code{NULL} or zero.}

\item{OwnLagMeans}{Prior expectation of own lag coefficients in reduced form.
See details.}

\item{flat}{Omit conventional uninformative prior on \code{Sigma}?}

\item{nonorm}{Do not normalize posterior to make it a proper prior}

\item{ic}{If non-null, do not use initial conditions from \code{ydata} in
forming the prior.  Use \code{ic} instead.}

\item{verbose}{If FALSE, return only the log marginal posterior density for `A0,lmd'.}
}
\value{
\itemize{
\item \code{mdd}: log Marginal posterior density for \code{A0}, \code{lmd}, with \verb{A+} integrated out.
\item \code{var}: Output of \code{svar()} for full sample, including dummy observations.
\item \code{varp}: output of \code{svar()} on prior dummy observations only.
}
\itemize{
\item \code{uts}: residuals for real data, as time series with dates if there are no
breaks, as a list of time series if there are breaks.
}
\itemize{
\item also: the input arguments
}
}
\description{
The posterior integrated over A+ (the right-hand side coefficients), conditional
on A0 and lmd.
}
\details{
The model is \deqn{A(L)y(t) = \varepsilon(t),} with \eqn{\varepsilon(t)}
having a diagonal covariance matrix that varies over time.  The variances change
at the dates in  \code{Tsigbreak}, and their relative sizes are in the \code{lmd} matrix.

The marginal density results are only available if the prior is proper, which
usually requires all or most of the prior parameters to be non-zero.  But
estimation with an improper prior that uses only some of the Minnesota prior
dummy observations is also possible.  For example \code{lambda=3}, \code{mu=1}, \code{tight=0},
\code{w=0},  (but \code{sig} non-zero) is a loose improper prior that insulates
against estimates that imply unlikely initial transients.   When the prior is
improper, \code{nonorm=TRUE} avoids some unnecessary calculations.

\code{OwnLagMeans} may be a single numeric value, the prior mean of the
first own lag coefficient in all equations.  If it is a numeric vector of
length m, it is the prior mean of the first m lag coefficients in all equations.
If it is a \code{nv} by \code{m} matrix, each row is the prior mean of the coefficients
on the first \code{m} lags in the corresponding equation.

The default \code{OwnLagMeans=c(1.25, -.25)} is close to the optimal second-order
univariate AR coefficients when the variable is a unit-averaged continuous time
Wiener process.  This works well for variables like GDP or investment, which
cumulate through time.  For data that are sampled rather than averaged, like
some financial or price data, \code{OwnLagMeans=1} is better. For near-i.i.d. series
like asset yields, `OwnLagMeans=0' might be better.

For more extensive discussion of the prior parameters see \link{MNpriorNotes.pdf}.

Though most commonly \code{ydata} is just a \code{mts} object, it can also be a list
of such objects.  The main use of this is to allow use of time series with
missing observations. The real data must be in blocks with increasing dates.

Data as a list also allows direct entry of dummy observations to function as
part of the prior, in addition to or in place of the dummy observations
generated by the parametric Minnesota prior. Such directly entered dummy
observations must come at the end of the list, and must be given the attribute
\code{dummy}, set to \code{TRUE}. In this case usually you will want \code{const=FALSE}.

A "training sample" prior can be implemented with a list element with attribute
\code{train} set to TRUE.  This is real data, with meaningful time series properties,
and must be the first part of the data set.  No shock variance breaks
(dates in \code{Tsigbrk}) can be in the training sample.

Changing or eliminating the \code{dummy} or
\code{train} attributes does not change the posterior distribution,
but it does affect the value of \code{mdd} and thus Bayes factors for model
comparison.
}
