% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysmat.R
\name{sysmat}
\alias{sysmat}
\title{System matrix}
\usage{
sysmat(By, Bx = NULL)
}
\arguments{
\item{By}{n by n by lags array of AR coefficients.  Output of \code{rfvar3} is in the right format.}

\item{Bx}{n by 1 vector of constant terms.}
}
\value{
The system matrix
}
\description{
Constructs the AR coefficient matrix for the stacked first-order AR
corresponding to the higher order VAR with autoregressive coefficient array
\code{By} and constant terms (if present) \code{Bx}.
}
\details{
If \code{Bx} is non-NULL, the system includes the constant in the AR dynamics.  That is,
the vector of ones whose coefficients are Bx is treated as one of the variables in the
system.
}
