% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimts.R
\name{trimts}
\alias{trimts}
\title{Trim mts object}
\usage{
trimts(y)
}
\arguments{
\item{y}{Multiple time series object to be trimmed.}
}
\description{
Remove initial and terminal observations that have NA values for any series.
}
