% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdecomp.R
\name{vdecomp}
\alias{vdecomp}
\title{VAR variance decomposition}
\usage{
vdecomp(resp)
}
\arguments{
\item{resp}{An array of dimension (number of variables) by (number of shocks) by (time horizon),
as produced, e.g., by \code{impulsdtrf}.}
}
\value{
\subsection{An array (\code{vdc} within the function) of the same dimension as \code{resp}.  \code{vdc[j, , ]} is a matrix in which each column represents the proportions of forecast error variance at horizon \code{j} accounted for by the various shocks.}{
}
}
\description{
Allocate explanatory power to shocks.
}
\details{
Converts an array of impulse responses  to an array of variance decompositions.
}
