% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2alm.R
\name{vec2alm}
\alias{vec2alm}
\title{Retrieve \code{A0} and lmd matrices from \code{x}}
\usage{
vec2alm(x, nv)
}
\value{
\describe{
\item{A}{\code{nv} by \code{nv} by \code{ndraw }A0\verb{ array\} \\item\{lmd\}\{}nv\code{by}nSig\code{by}ndraw\verb{ }lmd` array
}
\description{
\code{\link[=svarwrap]{svarwrap()}} takes \code{A0} and \code{lmd} collapsed into an \code{x} vector as argument.
This function reconstructs \code{A0} and \code{lmd} from \code{x}.
}
\details{
The function  handles both the case of a single \code{x} vector and
that of a matrix \code{x}, each row of which is expanded. The matrix case
applies to the output from \code{\link[=almdDraw]{almdDraw()}}
}
